+++ src/main/java/preisler/com/crazy_counter/workout/ExerciseDTO.java
@@ -8,7 +8,7 @@ import java.util.Date;
 @Setter
 @Getter
 public class ExerciseDTO {
-    private  int id;
+    private  Long id;
     private String name;
     private int reps;
     private int weight;
@@ -22,7 +22,7 @@ public class ExerciseDTO {
         this.weight = weight;
     }
 
-    public ExerciseDTO(int id, String name, int reps, int weight) {
+    public ExerciseDTO(Long id, String name, int reps, int weight) {
         this.id = id;
         this.name = name;
         this.reps = reps;
